+++ app/src/main/java/com/architectcoders/hotelapp/model/Constants.kt
