+++ app/src/main/java/com/architectcoders/hotelapp/model/Country.kt
@@ -1,4 +1,4 @@
-package com.architectcoders.hotelapp
+package com.architectcoders.hotelapp.model
 
 data class Country(
     val name: String,
