+++ app/src/main/java/com/architectcoders/hotelapp/model/DestinationResult.kt
@@ -1,4 +1,4 @@
-package com.architectcoders.hotelapp
+package com.architectcoders.hotelapp.model
 
 data class DestinationResult(
         val term: String,
