+++ app/src/main/java/com/architectcoders/hotelapp/model/HotelRetrofit.kt
@@ -1,5 +1,6 @@
-package com.architectcoders.hotelapp
+package com.architectcoders.hotelapp.model
 
+import com.architectcoders.hotelapp.API_BASE_URL
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
