+++ app/src/main/java/com/architectcoders/hotelapp/ui/HotelAdapter.kt
@@ -1,9 +1,11 @@
-package com.architectcoders.hotelapp
+package com.architectcoders.hotelapp.ui
 
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
+import com.architectcoders.hotelapp.model.Hotel
+import com.architectcoders.hotelapp.R
 import com.architectcoders.hotelapp.databinding.CustomListDestinationBinding
 import kotlin.properties.Delegates
